forces spark-based tests to run serially using semaphore
Issues with interop between scalatest and scoverage gradle plugins results in the scalatest arg that limits parallelism being ignored when tests are run via 'gradle build'.  Spark doesn't support multiple context being active in parallel, so we fallback on a semaphore to ensure that spark tests are always run serially.